pyproject.toml
[tool.poetry]
name = "conscience"
version = "0.1.0"
description = "Perform automated testing of TKinter GUI applications"
authors = ["Harry Keightley <harrykeightley@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
behave = "^1.2.6"
parse = "^1.20.1"
parse-type = "^0.6.2"
pillow = "^10.3.0"
six = "^1.16.0"
loguru = "^0.7.2"
setuptools = "^69.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
